Added taskbar and design components
