+++ Pages/Settings.tsx
@@ -1,0 +1,22 @@
+import {StyleSheet, Text, View} from 'react-native';
+import React, {Component} from 'react';
+import LinearGradient from 'react-native-linear-gradient';
+
+
+
+const Settings = () => {
+    return (
+        <Text>These are the settings</Text>
+    );
+};
+
+
+
+
+
+
+
+
+
+
+export default Settings;
