+++ package.json
@@ -10,13 +10,14 @@
     "test": "jest"
   },
   "dependencies": {
-    "@ant-design/react-native": "^5.1.0",
+    "@ant-design/react-native": "^4.2.0",
     "@react-native-async-storage/async-storage": "^1.23.1",
     "@react-native-community/segmented-control": "^2.2.2",
     "@react-native-community/slider": "^4.5.0",
     "@react-native-firebase/app": "^19.2.2",
     "@react-native-firebase/auth": "^19.2.2",
     "@react-native-firebase/storage": "^19.2.2",
+    "@react-navigation/bottom-tabs": "^6.5.20",
     "@react-navigation/core": "^6.4.14",
     "@react-navigation/native": "^6.1.17",
     "@react-navigation/native-stack": "^6.9.26",
@@ -29,6 +30,7 @@
     "react-native-linear-gradient": "^2.8.3",
     "react-native-safe-area-context": "^4.9.0",
     "react-native-screens": "^3.29.0",
+    "react-native-typography": "^1.4.1",
     "react-native-vector-icons": "^10.1.0",
     "validator": "^13.12.0"
   },
